Refactor code for modern C++ practices and readability
Refactored multiple classes and methods to improve code readability, maintainability, and safety by adopting modern C++ practices. Key changes include:

- Updated `CPegDoc::WriteShadowFile` to use `size_t` for type safety.
- Refactored `CPolyline` and `CPrim` classes for better structure and clarity.
- Replaced raw pointers with `std::vector` in `CPrimPolygon` and `CPrimPolyline` for safer memory management.
- Simplified logic in methods like `GoToNxtCtrlPt` and `SelUsingPoint`.
- Updated `CDrawHandler` to use dynamic `std::vector` for point storage, replacing fixed-size arrays.
- Reformatted namespaces and methods for consistency.
- Improved variable naming and removed redundant code.
- Added necessary includes for modern C++ features.

These changes enhance code safety, readability, and alignment with modern C++ standards.

+++ PegAEsysDoc.cpp
@@ -1264,7 +1264,7 @@ void CPegDoc::WriteShadowFile()
     {
         std::string shadowFilePath(app.GetShadowDir());
         shadowFilePath += GetTitle();
-        int extensionDot = shadowFilePath.find('.');
+        size_t extensionDot = shadowFilePath.find('.');
         if (extensionDot > 0)
         {
             CFileStatus fileStatus;
